plan to do inputNumber
